Fix the pinball tests.
